Port serial barcode scanner to current architecture
