--- docs/api/pylabrobot.capabilities.rst
@@ -1,367 +1,0 @@
-.. currentmodule:: pylabrobot.capabilities
-
-pylabrobot.capabilities package
-===============================
-
-Base classes for the capability system.
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    capability.Capability
-    capability.CapabilityBackend
-    capability.BackendParams
-
-
-Temperature Control
--------------------
-
-.. currentmodule:: pylabrobot.capabilities.temperature_controlling.temperature_controller
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    TemperatureController
-    TemperatureControllerBackend
-
-
-Shaking
--------
-
-.. currentmodule:: pylabrobot.capabilities.shaking.shaking
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Shaker
-    ShakerBackend
-
-
-Fan Control
------------
-
-.. currentmodule:: pylabrobot.capabilities.fan_control.fan_control
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Fan
-    FanBackend
-
-
-Humidity Control
-----------------
-
-.. currentmodule:: pylabrobot.capabilities.humidity_controlling.humidity_controller
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    HumidityController
-    HumidityControllerBackend
-
-
-Centrifuging
-------------
-
-.. currentmodule:: pylabrobot.capabilities.centrifuging.centrifuging
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Centrifuge
-    CentrifugeBackend
-
-
-Sealing
--------
-
-.. currentmodule:: pylabrobot.capabilities.sealing.sealing
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Sealer
-    SealerBackend
-
-
-Peeling
--------
-
-.. currentmodule:: pylabrobot.capabilities.peeling.peeling
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Peeler
-    PeelerBackend
-
-
-Loading Tray
-------------
-
-.. currentmodule:: pylabrobot.capabilities.loading_tray.loading_tray
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    LoadingTray
-
-.. currentmodule:: pylabrobot.capabilities.loading_tray.backend
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    LoadingTrayBackend
-
-
-Tilting
--------
-
-.. currentmodule:: pylabrobot.capabilities.tilting.tilting
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Tilter
-    TilterBackend
-
-
-Pumping
--------
-
-.. currentmodule:: pylabrobot.capabilities.pumping.pumping
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Pump
-    PumpBackend
-    PumpCalibration
-
-
-Weighing
---------
-
-.. currentmodule:: pylabrobot.capabilities.weighing.weighing
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Scale
-    ScaleBackend
-
-
-Barcode Scanning
-----------------
-
-.. currentmodule:: pylabrobot.capabilities.barcode_scanning.barcode_scanning
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    BarcodeScanner
-    BarcodeScannerBackend
-
-.. currentmodule:: pylabrobot.capabilities.barcode_scanning.serial
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    SerialBarcodeScanner
-    SerialBarcodeScannerDriver
-    SerialBarcodeScannerBackend
-
-
-Microscopy
-----------
-
-.. currentmodule:: pylabrobot.capabilities.microscopy.microscopy
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Microscopy
-    MicroscopyBackend
-
-
-Automated Retrieval
--------------------
-
-.. currentmodule:: pylabrobot.capabilities.automated_retrieval.automated_retrieval
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    AutomatedRetrieval
-    AutomatedRetrievalBackend
-
-
-Plate Reading - Absorbance
---------------------------
-
-.. currentmodule:: pylabrobot.capabilities.plate_reading.absorbance.absorbance
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Absorbance
-    AbsorbanceBackend
-
-
-Plate Reading - Fluorescence
------------------------------
-
-.. currentmodule:: pylabrobot.capabilities.plate_reading.fluorescence.fluorescence
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Fluorescence
-    FluorescenceBackend
-
-
-Plate Reading - Luminescence
------------------------------
-
-.. currentmodule:: pylabrobot.capabilities.plate_reading.luminescence.luminescence
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Luminescence
-    LuminescenceBackend
-
-
-Devices
--------
-
-.. currentmodule:: pylabrobot.thermo_fisher.multidrop_combi.multidrop_combi
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    MultidropCombi
-
-.. currentmodule:: pylabrobot.thermo_fisher.multidrop_combi.driver
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    MultidropCombiDriver
-
-
-Bulk Dispensing - Peristaltic
------------------------------
-
-.. currentmodule:: pylabrobot.capabilities.bulk_dispensers.peristaltic.peristaltic
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    PeristalticDispensing8
-
-.. currentmodule:: pylabrobot.capabilities.bulk_dispensers.peristaltic.backend
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    PeristalticDispensingBackend8
-
-
-Bulk Dispensing - Syringe
--------------------------
-
-.. currentmodule:: pylabrobot.capabilities.bulk_dispensers.syringe.syringe
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    SyringeDispensing8
-
-.. currentmodule:: pylabrobot.capabilities.bulk_dispensers.syringe.backend
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    SyringeDispensingBackend8
-
-
-Liquid Handling - PIP (Independent Channels)
---------------------------------------------
-
-.. currentmodule:: pylabrobot.capabilities.liquid_handling.pip
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    PIP
-    PIPBackend
-
-
-Liquid Handling - Head96 (96-Channel Head)
-------------------------------------------
-
-.. currentmodule:: pylabrobot.capabilities.liquid_handling.head96
-
-.. autosummary::
-  :toctree: _autosummary
-  :nosignatures:
-  :recursive:
-
-    Head96
-    Head96Backend
