+++ docs/api/pylabrobot.generic.rst
@@ -1,0 +1,14 @@
+.. currentmodule:: pylabrobot.generic
+
+Generic devices
+===============
+
+Drivers for devices that use common, configurable protocols rather than a manufacturer-specific
+protocol.
+
+.. autosummary::
+  :toctree: _autosummary
+  :nosignatures:
+  :recursive:
+
+    SerialBarcodeScanner
