+++ docs/user_guide/generic/index.md
@@ -1,0 +1,7 @@
+# Generic devices
+
+```{toctree}
+:maxdepth: 1
+
+line-barcode-scanner/hello-world
+```
