--- pylabrobot/capabilities/barcode_scanning/__init__.py
@@ -1,3 +1,0 @@
-from .backend import BarcodeScannerBackend, BarcodeScannerError
-from .barcode_scanning import BarcodeScanner
-from .serial import SerialBarcodeScanner, SerialBarcodeScannerBackend, SerialBarcodeScannerDriver
