+++ pylabrobot/generic/line_barcode_scanner/__init__.py
@@ -1,0 +1,1 @@
+from .serial import SerialBarcodeScanner
